requirements.txt
Django==5.0.4
djangorestframework==3.15.0
drf-spectacular==0.27.1
whitenoise==6.7.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-django==4.7.0
pytest-cov==4.1.0
coverage==7.3.2